requirements.txt
rembg
opencv-python
pillow
numpy